+++ Keas.Mvc/Services/IdentityService.cs
@@ -22,19 +22,80 @@ namespace Keas.Mvc.Services
         Task<User> GetByKerberos(string kerb);
         Task<string> BulkLoadPeople(string ppsCode, string teamslug);
         Task<PPSDepartmentResult> GetPpsDepartment(string ppsCode);
+
+        Task<(Person Person, int peopleCount)> GetOrCreateFromKerberos(string kerb, int teamId);
     }
 
     public class IdentityService : IIdentityService
     {
         private readonly AuthSettings _authSettings;
         private readonly ApplicationDbContext _context;
-        private readonly IPersonService _personService;
+    
 
-        public IdentityService(IOptions<AuthSettings> authSettings, ApplicationDbContext context, IPersonService personService)
+        public IdentityService(IOptions<AuthSettings> authSettings, ApplicationDbContext context)
         {
             _authSettings = authSettings.Value;
             _context = context;
-            _personService = personService;
+        }
+
+        public async Task<(Person Person, int peopleCount)> GetOrCreateFromKerberos(string kerb, int teamId)
+        {
+            var user = await _context.Users.Include(u => u.People).IgnoreQueryFilters().Where(u => u.Id == kerb).FirstOrDefaultAsync();
+            if (user == null)
+            {
+                //User does not exist, create, add to team
+                user = await GetByKerberos(kerb);
+                if (user != null)
+                {
+                    _context.Users.Add(user);
+                    var person = CreatePersonFromUser(user, teamId);
+                    _context.People.Add(person);
+                    await _context.SaveChangesAsync();
+                    return (person, 1);
+                }
+                else
+                {
+                    // No user found, return null
+                    return (null, 0);
+                }
+            }
+            else
+            {
+                //User found. Check if in team, add if not
+                var person = user.People.FirstOrDefault(p => p.TeamId == teamId);
+                if (person != null)
+                {
+                    //Person already in team, activate if needed
+                    if (!person.Active)
+                    {
+                        person.Active = true;
+                        await _context.SaveChangesAsync();
+                        return (person, 1);
+                    }
+                    return (person, 0);
+                }
+                else
+                {
+                    // Need to create person
+                    person = CreatePersonFromUser(user, teamId);
+                    _context.People.Add(person);
+                    await _context.SaveChangesAsync();
+                    return (person, 1);
+                }
+            }
+        }
+
+        private Person CreatePersonFromUser(User user, int teamId)
+        {
+            var person = new Person();
+            person.User = user;
+            person.FirstName = user.FirstName;
+            person.LastName = user.LastName;
+            person.Email = user.Email;
+            person.Active = true;
+            person.TeamId = teamId;
+            return person;
+
         }
 
         public async Task<User> GetByEmail(string email)
@@ -128,7 +189,7 @@ namespace Keas.Mvc.Services
 
                     if (kerbResults.ResponseData.Results.Length > 0 && contactResult.ResponseData.Results.Length > 0)
                     {
-                        var personResult = await _personService.GetOrCreateFromKerberos(kerbResults.ResponseData.Results[0].UserId, team.Id);
+                        var personResult = await GetOrCreateFromKerberos(kerbResults.ResponseData.Results[0].UserId, team.Id);
                         newpeople += personResult.peopleCount;
                     }
                 }
